third_party/blink/web_tests/external/wpt/css/css-align/blocks/align-content-block-003-expected.txt

This is a testharness.js-based test.
[FAIL] .test 5: last baseline
  assert_equals: \n<div class="test" style="align-content: last baseline" title="last baseline">\n    <div class="in-flow" data-offset-y="35"></div>\n    <div class="in-flow">\n      <span class="label">LAST BASELINE</span>\n      <span class="abspos" data-offset-y="20">ABS</span>\n      <span class="relpos" data-offset-y="20">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 35 but got 30
[FAIL] .test 9: unsafe center
  assert_equals: \n<div class="test" style="align-content: unsafe center" title="unsafe center">\n    <div class="in-flow" data-offset-y="25"></div>\n    <div class="in-flow">\n      <span class="label">UNSAFE CENTER</span>\n      <span class="abspos" data-offset-y="10">ABS</span>\n      <span class="relpos" data-offset-y="10">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 25 but got 35
[FAIL] .test 10: unsafe end
  assert_equals: \n<div class="test" style="align-content: unsafe end" title="unsafe end">\n    <div class="in-flow" data-offset-y="15"></div>\n    <div class="in-flow">\n      <span class="label">UNSAFE END</span>\n      <span class="abspos" data-offset-y="0">ABS</span>\n      <span class="relpos" data-offset-y="0">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 15 but got 35
Harness: the test ran to completion.
